% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{add_data_field_children_for_start_var}
\alias{add_data_field_children_for_start_var}
\title{Add DataField child nodes for start variable.}
\usage{
add_data_field_children_for_start_var(data_field, var_details_rows)
}
\arguments{
\item{data_field}{DataField node to attach child nodes.}

\item{var_details_rows}{Variable details rows associated with current variable.}
}
\value{
Updated DataField node.
}
\description{
Add DataField child nodes for start variable.
}
