% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{attach_apply_nodes}
\alias{attach_apply_nodes}
\title{Attach Apply nodes to a parent node.}
\usage{
attach_apply_nodes(var_details_rows, parent_node, db_name)
}
\arguments{
\item{var_details_rows}{Variable details rows associated with a variable.}

\item{parent_node}{An XML node.}

\item{db_name}{Database name.}
}
\value{
Updated parent node.
}
\description{
Attach Apply nodes to a parent node.
}
