% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{attach_cat_value_nodes_for_start_var}
\alias{attach_cat_value_nodes_for_start_var}
\title{Attach categorical value nodes to DataField node for start variable.}
\usage{
attach_cat_value_nodes_for_start_var(var_details_row, data_field)
}
\arguments{
\item{var_details_row}{Variable details sheet row.}

\item{data_field}{DataField node to attach Value nodes.}
}
\value{
Updated DataField node.
}
\description{
Attach categorical value nodes to DataField node for start variable.
}
