% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{attach_cont_value_nodes_for_start_var}
\alias{attach_cont_value_nodes_for_start_var}
\title{Attach continuous Value nodes for start variable.}
\usage{
attach_cont_value_nodes_for_start_var(var_details_row, data_field)
}
\arguments{
\item{var_details_row}{Variable details sheet row.}

\item{data_field}{DataField node to attach Value nodes.}
}
\value{
Updated DataField node.
}
\description{
Attach continuous Value nodes for start variable.
}
