% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{attach_range_value_nodes}
\alias{attach_range_value_nodes}
\title{Attach Value nodes to DataField node. Used when `recFrom` has a value range.}
\usage{
attach_range_value_nodes(var_details_row, data_field)
}
\arguments{
\item{var_details_row}{Variable details sheet row.}

\item{data_field}{DataField node to attach Value nodes.}
}
\value{
Updated DataField node.
}
\description{
Attach Value nodes to DataField node. Used when `recFrom` has a value range.
}
