% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{build_data_field_for_start_var}
\alias{build_data_field_for_start_var}
\title{Build DataField node for start variable.}
\usage{
build_data_field_for_start_var(var_name, var_details_rows)
}
\arguments{
\item{var_name}{Variable name.}

\item{var_details_rows}{All variable details rows for the `var_name` variable.}
}
\value{
DataField node with optype and dataType according to `fromType`.
}
\description{
Build DataField node for start variable.
}
