% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{build_data_field_for_var}
\alias{build_data_field_for_var}
\title{Build DataField node for variable.}
\usage{
build_data_field_for_var(var_name, vars_sheet)
}
\arguments{
\item{var_name}{Variable name.}

\item{vars_sheet}{Variable sheet data frame.}
}
\value{
DataField node for variable.
}
\description{
Build DataField node for variable.
}
