% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{build_derived_field_value_node}
\alias{build_derived_field_value_node}
\title{Build Value node for DerivedField node.}
\usage{
build_derived_field_value_node(var_details_row)
}
\arguments{
\item{var_details_row}{Variable details sheet row.}
}
\value{
Value node.
}
\description{
Build Value node for DerivedField node.
}
