% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{build_missing_const_node}
\alias{build_missing_const_node}
\title{Build Constant node for a missing value for a variable.}
\usage{
build_missing_const_node(var_details_row)
}
\arguments{
\item{var_details_row}{Variable details sheet row.}
}
\value{
Constant node.
}
\description{
Build Constant node for a missing value for a variable.
}
