% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{build_ranged_derived_field_apply_node}
\alias{build_ranged_derived_field_apply_node}
\title{Build Apply node with interval nodes for DerivedField node.}
\usage{
build_ranged_derived_field_apply_node(var_details_row, db_name)
}
\arguments{
\item{var_details_row}{Variable details sheet row.}

\item{db_name}{Database name.}
}
\value{
Apply node with intervals for DerivedField node.
}
\description{
Build Apply node with interval nodes for DerivedField node.
}
