% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{build_trans_dict}
\alias{build_trans_dict}
\title{Build a TransformationDictionary node.}
\usage{
build_trans_dict(vars_sheet, var_details_sheet, var_names, db_name)
}
\arguments{
\item{vars_sheet}{Variable sheet data frame.}

\item{var_details_sheet}{Variable details sheet data frame.}

\item{var_names}{Vector of variable names.}

\item{db_name}{Database name.}
}
\value{
TransformationDictionary node.
}
\description{
Build a TransformationDictionary node.
}
