% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{build_variable_field_ref_node}
\alias{build_variable_field_ref_node}
\title{Build FieldRef node for variable.}
\usage{
build_variable_field_ref_node(var_details_row, db_name)
}
\arguments{
\item{var_details_row}{Variable details sheet row.}

\item{db_name}{Database name.}
}
\value{
FieldRef node.
}
\description{
Build FieldRef node for variable.
}
