% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_id_row}
\alias{create_id_row}
\title{ID role creation}
\usage{
create_id_row(data, id_role_name, database_name, variables)
}
\arguments{
\item{data}{the data that the ID role row is created for}

\item{id_role_name}{name for the role that ID is created from}

\item{database_name}{the name of the database}

\item{variables}{variables sheet containing variable information}
}
\value{
data with the ID row attached
}
\description{
Creates ID row for rec_with_table
}
