% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_label_list_element}
\alias{create_label_list_element}
\title{Create label list element}
\usage{
create_label_list_element(variable_rows)
}
\arguments{
\item{variable_rows}{all variable details rows containing 1 variable information}
}
\value{
a list containing labels for the passed variable
}
\description{
A data labeling utility function for creating individual variable labels
}
