% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_der_function.R
\name{example_der_fun}
\alias{example_der_fun}
\title{example_der_fun caluclates chol*bili}
\usage{
example_der_fun(chol, bili)
}
\arguments{
\item{chol}{the row value for chol}

\item{bili}{the row value for bili}
}
\description{
example_der_fun caluclates chol*bili
}
