% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_margin_closure}
\alias{get_margin_closure}
\title{Get closure type for a margin.}
\usage{
get_margin_closure(chars)
}
\arguments{
\item{chars}{Character vector.}
}
\value{
Closure type.
}
\description{
Get closure type for a margin.
}
