% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_margins}
\alias{get_margins}
\title{Extract margins from character vector.}
\usage{
get_margins(chars)
}
\arguments{
\item{chars}{Character vector.}
}
\value{
Margins as character vector.
}
\description{
Extract margins from character vector.
}
