% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_start_var_name}
\alias{get_start_var_name}
\title{Get variable name from variableStart using database name.}
\usage{
get_start_var_name(var_details_row, db_name)
}
\arguments{
\item{var_details_row}{A variable details row.}

\item{db_name}{Name of database to extract from.}
}
\value{
character The name of the start variable.
}
\description{
Get variable name from variableStart using database name.
}
