% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_var_details_row_indices}
\alias{get_var_details_row_indices}
\title{Get all variable details row indices for a variable.}
\usage{
get_var_details_row_indices(var_details_sheet, var_name)
}
\arguments{
\item{var_details_sheet}{A data frame representing a variable details sheet.}

\item{var_name}{Variable name.}
}
\value{
All variable details row indices for a variable.
}
\description{
Get all variable details row indices for a variable.
}
