% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_var_details_rows}
\alias{get_var_details_rows}
\title{Get all variable details rows for a variable and database combination.}
\usage{
get_var_details_rows(var_details_sheet, var_name, db_name)
}
\arguments{
\item{var_details_sheet}{A data frame representing a variable details sheet.}

\item{var_name}{Variable name.}

\item{db_name}{Database name.}
}
\value{
All variable details rows for the variable and database combination.
}
\description{
Get all variable details rows for a variable and database combination.
}
