% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_var_sheet_row}
\alias{get_var_sheet_row}
\title{Get variable row from variable sheet.}
\usage{
get_var_sheet_row(var_name, vars_sheet)
}
\arguments{
\item{var_name}{Variable name.}

\item{vars_sheet}{Variable sheet data frame.}
}
\value{
Variable row.
}
\description{
Get variable row from variable sheet.
}
