% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_variable_type_data_type}
\alias{get_variable_type_data_type}
\title{Get data type for variable type.}
\usage{
get_variable_type_data_type(var_details_rows, var_type, is_start_var)
}
\arguments{
\item{var_details_rows}{All variable details rows for the variable.}

\item{var_type}{Variable type}

\item{is_start_var}{boolean if the passed variable is variable start}
}
\value{
`var_type` data type.
}
\description{
Get data type for variable type.
}
