% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode-with-table.R
\name{is_derived_var}
\alias{is_derived_var}
\title{Whether a variable in a variables details sheet is a derived variable}
\usage{
is_derived_var(variable_details_row)
}
\arguments{
\item{variable_details_row}{A data frame with a single row which will be
checked}
}
\value{
A boolean
}
\description{
Whether a variable in a variables details sheet is a derived variable
}
