% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode-with-table.R
\name{is_equal}
\alias{is_equal}
\title{Checks whether two values are equal including NA}
\usage{
is_equal(v1, v2)
}
\arguments{
\item{v1}{variable 1}

\item{v2}{variable 2}
}
\value{
boolean value of whether or not v1 and v2 are equal
}
\description{
Compared to the base "==" operator in R, this function returns
true if the two values are NA
whereas the base "==" operator returns NA
}
\examples{
is_equal(1,2)
# FALSE

is_equal(1,1)
# TRUE

1==NA
# NA

is_equal(1,NA)
# FALSE

NA==NA
# NA

is_equal(NA,NA)
# TRUE
}
