% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_left_open}
\alias{is_left_open}
\title{Extract margins from character vector.}
\usage{
is_left_open(chars)
}
\arguments{
\item{chars}{Character vector.}
}
\value{
Whether the left endpoint of an interval is open.
}
\description{
Extract margins from character vector.
}
