% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_numeric}
\alias{is_numeric}
\title{Check if a character object can be converted to a number.}
\usage{
is_numeric(chars)
}
\arguments{
\item{chars}{Character object.}
}
\value{
Whether `chars` can be converted to a numeric value.
}
\description{
Check if a character object can be converted to a number.
}
