% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_rec_from_range}
\alias{is_rec_from_range}
\title{Check if recFrom is a range for a variable details row.}
\usage{
is_rec_from_range(var_details_row)
}
\arguments{
\item{var_details_row}{Variable details sheet row.}
}
\value{
Whether recFrom is a range.
}
\description{
Check if recFrom is a range for a variable details row.
}
