% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-start.R
\name{is_table_feeder_var}
\alias{is_table_feeder_var}
\title{Checks whether a variable start is a table}
\usage{
is_table_feeder_var(feeder_var)
}
\arguments{
\item{feeder_var}{string The variable start to check}
}
\value{
bool
}
\description{
Checks whether a variable start is a table
}
