% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbc_metadata}
\alias{pbc_metadata}
\title{Metadata for the pbc dataset using the DCIM standard}
\format{
A list containing DCMI metadata:
\describe{
\item{title}{title}
\item{creator}{creator}
\item{subject}{subject}
\item{description}{description}
\item{publisher}{publisher}
\item{date}{date}
\item{type}{type}
\item{format}{format}
\item{identifier}{identifier}
\item{source}{source}
\item{language}{language}
\item{rights}{rights}
\item{references}{references}
}
}
\usage{
pbc_metadata
}
\description{
Metadata for the pbc dataset using the DCIM standard
}
\keyword{datasets}
