% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbc_variables}
\alias{pbc_variables}
\title{Variables sheet for the pbc dataset}
\format{
A data frame with 24 rows and 11 columns:
\describe{
\item{variable}{variable name}
\item{label}{variable label}
\item{labelLong}{variable label long}
\item{subject}{subject}
\item{section}{section}
\item{variableType}{variable type}
\item{databaseStart}{database start}
\item{units}{units}
\item{variableStart}{variable start}
\item{notes}{logical indicating presence of notes}
\item{description}{logical indicating presence of description}
}
}
\usage{
pbc_variables
}
\description{
Variables sheet for the pbc dataset
}
\keyword{datasets}
