% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_vars_by_role}
\alias{select_vars_by_role}
\title{Vars selected by role}
\usage{
select_vars_by_role(roles, variables)
}
\arguments{
\item{roles}{a vector containing a single or multiple roles to match by}

\item{variables}{the variables sheet containing variable info}
}
\value{
a vector containing the variable names that match the passed roles
}
\description{
Selects variables from variables sheet based on passed roles
}
