% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tester_variable_details}
\alias{tester_variable_details}
\title{Example variable details sheet for vignettes}
\format{
A data frame with 69 rows and 16 columns:
\describe{
\item{variable}{variable name}
\item{dummyVariable}{dummy variable name}
\item{typeEnd}{end type}
\item{databaseStart}{database start}
\item{variableStart}{variable start}
\item{typeStart}{start type}
\item{recEnd}{record end}
\item{recStart}{record start}
\item{catLabel}{category label}
\item{catLabelLong}{category long label}
\item{numValidCat}{number of valid categories (numeric)}
\item{units}{logical indicating presence of units}
\item{notes}{logical indicating presence of notes}
\item{catStartLabel}{category start label}
\item{variableStartShortLabel}{variable start short label}
\item{variableStartLabel}{variable start label}
}
}
\usage{
tester_variable_details
}
\description{
Example variable details sheet for vignettes
}
\keyword{datasets}
