% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tester_variables}
\alias{tester_variables}
\title{Example variables sheet for vignettes}
\format{
A data frame with 24 rows and 11 columns:
\describe{
\item{variable}{variable name}
\item{label}{variable label}
\item{labelLong}{variable label long}
\item{subject}{subject}
\item{section}{section}
\item{variableType}{variable type}
\item{databaseStart}{database start}
\item{units}{units}
\item{variableStart}{variable start}
\item{notes}{logical indicating presence of notes}
\item{description}{logical indicating presence of description}
}
}
\usage{
tester_variables
}
\description{
Example variables sheet for vignettes
}
\keyword{datasets}
