% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmml.R
\name{xflow_to_pmml}
\alias{xflow_to_pmml}
\title{Creates a PMML document from an xFlow document.}
\usage{
xflow_to_pmml(var_details_sheet, vars_sheet, db_name, vars_to_convert = NULL)
}
\arguments{
\item{var_details_sheet}{A data frame representing a variable details sheet.}

\item{vars_sheet}{A data frame representing a variables sheet.}

\item{db_name}{A string containing the name of the database that holds
the start variables. Should match up with one of the databases in the
databaseStart column.}

\item{vars_to_convert}{A vector of strings containing the names of variables
from the variable column in the variable details sheet that should be
converted to PMML. Passing in an empty vector will convert all the variables.}
}
\value{
A PMML document.
}
\description{
Creates a PMML document from an xFlow document.
}
\examples{

}
