% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-start.R
\name{get_table_name}
\alias{get_table_name}
\title{Returns the name of the table for a table start variable}
\usage{
get_table_name(table_feeder_var)
}
\arguments{
\item{table_feeder_var}{string The table variable start}
}
\value{
string
}
\description{
Returns the name of the table for a table start variable
}
